requirements.txt
aiohttp==3.7.4.post0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
authy==2.2.6
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
clx-sdk-xms==1.0.0
colorama==0.4.3
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-messages==0.6.0
djangorestframework==3.11.1
geoip2==4.1.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
iso8601==0.1.14
isort==5.4.2
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.1
pandas==1.1.1
phonenumbers==8.12.19
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
sinchsms==1.0.4
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
twilio==6.54.0
typing-extensions==3.7.4.3
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
